Refactor print file builder code out into it's own module Add a separate configurable logging level for pika Update README
